Move v5 Radio to "legacy" export
